refactor: update modular project structure
