+++ src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { UserController } from './controllers/user/user.controller';
-import { AlbumController } from './controllers/album/album.controller';
-import { ArtistController } from './controllers/artist/artist.controller';
-import { FavsController } from './controllers/favs/favs.controller';
-import { TrackController } from './controllers/track/track.controller';
+import { UserController } from './user/user.controller';
+import { AlbumController } from './album/album.controller';
+import { ArtistController } from './artist/artist.controller';
+import { FavsController } from './favs/favs.controller';
+import { TrackController } from './track/track.controller';
 
 @Module({
   imports: [],
